+++ public/src/admin/settings/homepage.js
@@ -1,9 +1,9 @@
 define('admin/settings/homepage', ['admin/settings'], function () {
 	function toggleCustomRoute() {
-		if ($('[data-field="homePageRoute"]').val()) {
-			$('#homePageCustom').hide();
-		} else {
+		if ($('[data-field="homePageRoute"]').val() === 'custom') {
 			$('#homePageCustom').show();
+		} else {
+			$('#homePageCustom').hide();
 		}
 	}