+++ public/src/client/flags/detail.js
@@ -118,8 +118,6 @@ define('forum/flags/detail', ['forum/flags/list', 'components', 'translator', 'b
 					break;
 			}
 		});
-
-		FlagsList.enableFilterForm();
 	};
 
 	function postAction(action, pid, tid) {