+++ public/src/client/topic/delete-posts.js
@@ -49,7 +49,6 @@ define('forum/topic/delete-posts', ['components', 'postSelect'], function (compo
 	function deletePosts(btn, command) {
 		btn.attr('disabled', true);
 		socket.emit(command, {
-			tid: ajaxify.data.tid,
 			pids: postSelect.pids,
 		}, function (err) {
 			btn.removeAttr('disabled');