+++ public/src/client/topic/move.js
@@ -1,4 +1,7 @@
-define('forum/topic/move', ['categorySelector'], function (categorySelector) {
+'use strict';
+
+
+define('forum/topic/move', ['categorySelector', 'alerts'], function (categorySelector, alerts) {
 	var Move = {};
 	var modal;
 	var selectedCategory;
@@ -43,7 +46,22 @@ define('forum/topic/move', ['categorySelector'], function (categorySelector) {
 		if (!commitEl.prop('disabled') && selectedCategory && selectedCategory.cid) {
 			commitEl.prop('disabled', true);
 
-			moveTopics();
+			modal.modal('hide');
+
+			alerts.alert({
+				alert_id: 'tids_move_' + Move.tids.join(';'),
+				title: '[[topic:thread_tools.move]]',
+				message: '[[topic:topic_move_success, ' + selectedCategory.name + ']]',
+				type: 'success',
+				timeout: 5000,
+				timeoutfn: function () {
+					moveTopics();
+				},
+				clickfn: function (alert, params) {
+					delete params.timeoutfn;
+					app.alertSuccess('[[topic:topic_move_undone]]');
+				},
+			});
 		}
 	}