+++ public/src/modules/alerts.js
@@ -1,58 +1,122 @@
-import benchpress from 'benchpressjs';
-import translator from './translator';
-import components from './components';
+'use strict';
 
-export function alert(params) {
-	params.alert_id = 'alert_button_' + (params.alert_id ? params.alert_id : new Date().getTime());
-	params.title = params.title ? params.title.trim() || '' : '';
-	params.message = params.message ? params.message.trim() : '';
-	params.type = params.type || 'light';
 
-	var alert = $('#' + params.alert_id);
-	if (alert.length) {
-		updateAlert(alert, params);
-	} else {
-		createNew(params);
-	}
-}
+define('alerts', ['translator', 'components', 'benchpress'], function (translator, components, Benchpress) {
+	var module = {};
 
-export function remove(id) {
-	$('#alert_button_' + id).remove();
-}
+	module.alert = function (params) {
+		params.alert_id = 'alert_button_' + (params.alert_id ? params.alert_id : new Date().getTime());
+		params.title = params.title ? params.title.trim() || '' : '';
+		params.message = params.message ? params.message.trim() : '';
+		params.type = params.type || 'info';
 
-function createNew(params) {
-	benchpress.parse('partials/alert', params, function (alertTpl) {
-		translator.translate(alertTpl, function (translatedHTML) {
-			var alert = $('#' + params.alert_id);
-			if (alert.length) {
-				return updateAlert(alert, params);
-			}
-			alert = $(translatedHTML);
+		var alert = $('#' + params.alert_id);
+		if (alert.length) {
+			updateAlert(alert, params);
+		} else {
+			createNew(params);
+		}
+	};
 
-			components.get('toaster/tray').prepend(alert);
-			if (typeof params.closefn === 'function') {
-				alert.find('button').on('click', function () {
-					params.closefn();
-					return false;
-				});
-			}
+	function createNew(params) {
+		Benchpress.parse('alert', params, function (alertTpl) {
+			translator.translate(alertTpl, function (translatedHTML) {
+				var alert = $('#' + params.alert_id);
+				if (alert.length) {
+					return updateAlert(alert, params);
+				}
+				alert = $(translatedHTML);
+				alert.fadeIn(200);
 
-			if (typeof params.clickfn === 'function') {
-				alert
-					.addClass('pointer')
-					.on('click', function (e) {
-						if (!$(e.target).is('.close')) {
-							params.clickfn();
-						}
+				components.get('toaster/tray').prepend(alert);
+
+				if (typeof params.closefn === 'function') {
+					alert.find('button').on('click', function () {
+						params.closefn();
+						fadeOut(alert);
+						return false;
 					});
-			}
+				}
 
-			alert.toast('show');
+				if (params.timeout) {
+					startTimeout(alert, params);
+				}
 
-			$(window).trigger('action:alert.new', { alert: alert, params: params });
+				if (typeof params.clickfn === 'function') {
+					alert
+						.addClass('pointer')
+						.on('click', function (e) {
+							if (!$(e.target).is('.close')) {
+								params.clickfn(alert, params);
+							}
+							fadeOut(alert);
+						});
+				}
+
+				$(window).trigger('action:alert.new', { alert: alert, params: params });
+			});
 		});
-	});
-}
+	}
+
+	module.remove = function (id) {
+		$('#alert_button_' + id).remove();
+	};
+
+	function updateAlert(alert, params) {
+		alert.find('strong').html(params.title);
+		alert.find('p').html(params.message);
+		alert.attr('class', 'alert alert-dismissable alert-' + params.type + ' clearfix');
+
+		clearTimeout(parseInt(alert.attr('timeoutId'), 10));
+		if (params.timeout) {
+			startTimeout(alert, params);
+		}
+
+		alert.children().fadeOut(100);
+		translator.translate(alert.html(), function (translatedHTML) {
+			alert.children().fadeIn(100);
+			alert.html(translatedHTML);
+			$(window).trigger('action:alert.update', { alert: alert, params: params });
+		});
+
+		// Handle changes in the clickfn
+		alert.off('click').removeClass('pointer');
+		if (typeof params.clickfn === 'function') {
+			alert
+				.addClass('pointer')
+				.on('click', function (e) {
+					if (!$(e.target).is('.close')) {
+						params.clickfn();
+					}
+					fadeOut(alert);
+				});
+		}
+	}
+
+	function fadeOut(alert) {
+		alert.fadeOut(500, function () {
+			$(this).remove();
+		});
+	}
+
+	function startTimeout(alert, params) {
+		var timeout = params.timeout;
+
+		var timeoutId = setTimeout(function () {
+			fadeOut(alert);
+
+			if (typeof params.timeoutfn === 'function') {
+				params.timeoutfn(alert, params);
+			}
+		}, timeout);
+
+		alert.attr('timeoutId', timeoutId);
+
+		// Reset and start animation
+		alert.css('transition-property', 'none');
+		alert.removeClass('animate');
+	}
+});
 
 function updateAlert(alert, params) {
 	alert.find('strong').html(params.title);