+++ src/meta/templates.js
@@ -8,8 +8,6 @@ const rimraf = require('rimraf');
 const winston = require('winston');
 const path = require('path');
 const fs = require('fs');
-const fsReadFile = util.promisify(fs.readFile);
-const fsWriteFile = util.promisify(fs.writeFile);
 
 const nconf = require('nconf');
 const _ = require('lodash');
@@ -34,7 +32,7 @@ async function processImports(paths, templatePath, source) {
 
 	var partial = matches[1];
 	if (paths[partial] && templatePath !== partial) {
-		const partialSource = await fsReadFile(paths[partial], 'utf8');
+		const partialSource = await fs.promises.readFile(paths[partial], 'utf8');
 		source = source.replace(regex, partialSource);
 		return await processImports(paths, templatePath, source);
 	}
@@ -118,7 +116,7 @@ async function compileTemplate(filename, source) {
 	const compiled = await Benchpress.precompile(source, {
 		minify: process.env.NODE_ENV !== 'development',
 	});
-	return await fsWriteFile(path.join(viewsPath, filename.replace(/\.tpl$/, '.js')), compiled);
+	return await fs.promises.writeFile(path.join(viewsPath, filename.replace(/\.tpl$/, '.js')), compiled);
 }
 Templates.compileTemplate = compileTemplate;
 
@@ -141,16 +139,15 @@ async function compile() {
 
 	await async.eachLimit(names, 4, async (name) => {
 		const filePath = files[name];
-		let imported = await fsReadFile(filePath, 'utf8');
+		let imported = await fs.promises.readFile(filePath, 'utf8');
 		imported = await processImports(files, name, imported);
 
-		const compiled = await Benchpress.precompile(imported, { minify: process.env.NODE_ENV !== 'development' });
+		await mkdirp(path.join(viewsPath, path.dirname(name)));
 
-		await Promise.all([
-			fsWriteFile(path.join(viewsPath, name), imported),
-			fsWriteFile(path.join(viewsPath, name.replace(/\.tpl$/, '.js')), compiled),
-		]);
-	});
+		await fs.promises.writeFile(path.join(viewsPath, name), imported);
+		const compiled = await Benchpress.precompile(imported, { minify: process.env.NODE_ENV !== 'development' });
+		await fs.promises.writeFile(path.join(viewsPath, name.replace(/\.tpl$/, '.js')), compiled);
+	}));
 
 	winston.verbose('[meta/templates] Successfully compiled templates.');
 }