+++ src/plugins/index.js
@@ -11,8 +11,6 @@ const util = require('util');
 const user = require('../user');
 const posts = require('../posts');
 
-const readdirAsync = util.promisify(fs.readdir);
-
 const Plugins = module.exports;
 
 require('./install')(Plugins);
@@ -250,7 +248,7 @@ Plugins.normalise = async function (apiReturn) {
 Plugins.nodeModulesPath = path.join(__dirname, '../../node_modules');
 
 Plugins.showInstalled = async function () {
-	const dirs = await readdirAsync(Plugins.nodeModulesPath);
+	const dirs = await fs.promises.readdir(Plugins.nodeModulesPath);
 
 	let pluginPaths = await findNodeBBModules(dirs);
 	pluginPaths = pluginPaths.map(dir => path.join(Plugins.nodeModulesPath, dir));